feat(console): User can filter per timeframe on analytics dashboard at API level
