+++ gravitee-apim-console-webui/src/management/api/api-traffic-v4/analytics/components/api-analytics-proxy-filter-bar/api-analytics-proxy-filter-bar.component.html
@@ -15,52 +15,4 @@
     limitations under the License.
 
 -->
-<form [formGroup]="form" class="form">
-  <mat-form-field>
-    <mat-label>Timeframe</mat-label>
-    <mat-select formControlName="period">
-      @for (timeFrame of timeFrames; track timeFrame.id) {
-        <mat-option [value]="timeFrame.id">{{ timeFrame.label }}</mat-option>
-      }
-    </mat-select>
-  </mat-form-field>
-
-  @if (form.controls.period.value === customPeriod) {
-    <mat-form-field class="custom-date">
-      <mat-label>From</mat-label>
-      <input matInput [owlDateTime]="fromDateTimePicker" [max]="nowDate" formControlName="from" />
-      <mat-icon [owlDateTimeTrigger]="fromDateTimePicker" matSuffix svgIcon="gio:nav-arrow-down"></mat-icon>
-      <owl-date-time #fromDateTimePicker></owl-date-time>
-    </mat-form-field>
-
-    <mat-form-field class="custom-date">
-      <mat-label>To</mat-label>
-      <input matInput [owlDateTime]="toDateTimePicker" [min]="minDate" [max]="nowDate" formControlName="to" />
-      <mat-icon [owlDateTimeTrigger]="toDateTimePicker" matSuffix svgIcon="gio:nav-arrow-down"></mat-icon>
-      <owl-date-time #toDateTimePicker></owl-date-time>
-      @if (form.hasError('dateRange')) {
-        <mat-error>To date cannot be earlier than From date</mat-error>
-      }
-    </mat-form-field>
-
-    <button
-      data-testid="apply-button"
-      [disabled]="form.invalid || !form.controls.from.value || !form.controls.to.value"
-      mat-stroked-button
-      aria-label="Apply"
-      (click)="applyCustomTimeframe()"
-    >
-      Apply
-    </button>
-  } @else {
-    <button data-testid="refresh-button" mat-stroked-button aria-label="Refresh data" (click)="refreshFilters()">Refresh data</button>
-  }
-
-  <div class="errors">
-    @if (form.hasError('dateRange')) {
-      <mat-error>
-        <span>Error: "To" date cannot be earlier than "From" date</span>
-      </mat-error>
-    }
-  </div>
-</form>
+<gio-timeframe-widget [activeFilters]="activeFilters()" (filtersChange)="filtersChange.emit($event)" (refresh)="refresh.emit()" />
