+++ gravitee-apim-console-webui/src/shared/components/gio-timeframe-widget/gio-timeframe-widget.component.html
@@ -1,0 +1,72 @@
+<!--
+
+    Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+            http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<form [formGroup]="form" class="form">
+  <mat-form-field>
+    <mat-label>Timeframe</mat-label>
+    <mat-select formControlName="period">
+      @for (timeFrame of timeFrames; track timeFrame.id) {
+        <mat-option [value]="timeFrame.id">
+          <div class="option-radio-container">
+            <mat-radio-button [checked]="form.get('period')?.value === timeFrame.id" [value]="timeFrame.id" disableRipple>
+              {{ timeFrame.label }}
+            </mat-radio-button>
+          </div>
+        </mat-option>
+      }
+    </mat-select>
+  </mat-form-field>
+
+  @if (form.controls.period.value === customPeriod) {
+    <mat-form-field class="custom-date">
+      <mat-label>From</mat-label>
+      <input matInput [owlDateTime]="fromDateTimePicker" [max]="nowDate" formControlName="from" />
+      <mat-icon [owlDateTimeTrigger]="fromDateTimePicker" matSuffix svgIcon="gio:calendar"></mat-icon>
+      <owl-date-time #fromDateTimePicker></owl-date-time>
+    </mat-form-field>
+
+    <mat-form-field class="custom-date">
+      <mat-label>To</mat-label>
+      <input matInput [owlDateTime]="toDateTimePicker" [min]="minDate" [max]="nowDate" formControlName="to" />
+      <mat-icon [owlDateTimeTrigger]="toDateTimePicker" matSuffix svgIcon="gio:calendar"></mat-icon>
+      <owl-date-time #toDateTimePicker></owl-date-time>
+      @if (form.hasError('dateRange')) {
+        <mat-error>To date cannot be earlier than From date</mat-error>
+      }
+    </mat-form-field>
+
+    <button
+      data-testid="apply-button"
+      [disabled]="form.invalid || !form.controls.from.value || !form.controls.to.value"
+      mat-stroked-button
+      aria-label="Apply"
+      (click)="applyCustomTimeframe()"
+    >
+      Apply
+    </button>
+  } @else {
+    <button data-testid="refresh-button" mat-stroked-button aria-label="Refresh data" (click)="refreshFilters()">Refresh data</button>
+  }
+
+  <div class="errors">
+    @if (form.hasError('dateRange')) {
+      <mat-error>
+        <span>Error: "To" date cannot be earlier than "From" date</span>
+      </mat-error>
+    }
+  </div>
+</form>
